Change HoursDto to use int for hours

+++ Tempus.API/Controllers/Dto/HoursDto.cs
@@ -6,7 +6,7 @@ namespace Tempus.API.Controllers.Dto
     public class HoursDto
     {
         public Int64 Date { get; set; }
-        public float Hours { get; set; }
+        public Int16 Hours { get; set; }
         public string Project { get; set; }
         public string Notes { get; set; }
     }
